move changes to wt_verify.py
