+++ tools/wt_verify.py
@@ -47,13 +47,15 @@ def print_output(output):
     Print output for parsing script
     """
     str = "{"
-    for checkpoint in output:
+    for i, checkpoint in enumerate(output):
+        if i:
+            str += ","
         str += "\n\t" + checkpoint + ": {"
-        for count, key in enumerate(output[checkpoint]):
+        for j, key in enumerate(output[checkpoint]):
             value = output[checkpoint][key]
-            if count:
-                str += ", "
-            str += "\n\t\t" + key + ": {\n\t\t\t\"metadata\": " + json.dumps(value["metadata"]) 
+            if j:
+                str += ","
+            str += "\n\t\t" + key + ": {\n\t\t\t" + json.dumps(value)
             if "data" in value:
                 str += "\n\t\t\t\"data\": " + json.dumps(value["data"]) 
             if "hs_modify" in value:
@@ -95,6 +97,22 @@ def string_to_iterable(line):
     return dict
 
 
+def parse_node(line, output, checkpoint_name, cur_node, cur_node_id):
+    line = line[len("- "):-1]
+    page_type = ([line.split(": ")[0]] + line.split(": ")[1].split())[-1]
+    if page_type == "root":
+        cur_node = {}
+        cur_node_id = is_int(line.split(": ")[0])
+        cur_node["page_type"] = [page_type]
+    elif page_type == "internal" or page_type == "leaf":
+        output[checkpoint_name][cur_node_id] = cur_node 
+        cur_node_id = is_int(line.split(": ")[0])
+        cur_node = {}
+    else:
+        pass
+    return cur_node_id
+
+
 def parse_output(file_path):
     """
     Parse the output file of dump_pages
@@ -102,36 +120,23 @@ def parse_output(file_path):
     separator = "=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=\n"
     f = open(file_path, "r")
     output = {}
-    checkpoint = {}
     line = f.readline() # separator
-    cur_node = {"metadata": {}}
+    cur_node = {}
     cur_node_id = None
-    page_type = None
 
     while line:
         if line == separator:
             line = f.readline() # checkpoint
             checkpoint_name = line.split(", ")[-1].split(": ")[-1][:-1]
-            output[checkpoint_name] = checkpoint
+            output[checkpoint_name] = {}
         while line != separator and line:
             if line[0:2] == "- ": # start of a new node
-                line = line[len("- "):-1]
-                page_info = [line.split(": ")[0]] + line.split(": ")[1].split()
-                page_type = page_info[-1]
-                if page_info[0] == "Root":
-                    cur_node_id = page_info[0]
-                    cur_node["metadata"]["store_type"] = page_info[1]
-                elif page_type == "internal" or page_type == "leaf":
-                    checkpoint[cur_node_id] = cur_node 
-                    cur_node_id = is_int(line.split(": ")[0])
-                    cur_node = {"metadata": {}}
-                else:
-                    pass
+                cur_node_id = parse_node(line, output, checkpoint_name, cur_node, cur_node_id)
             elif line[0:3] == "\t> ": # metadata for new node
-                cur_node["metadata"].update(string_to_iterable(line[len("\t> "):-1]))
+                cur_node.update(string_to_iterable(line[len("\t> "):-1]))
             line = f.readline() 
         if cur_node_id is not None:
-            checkpoint[cur_node_id] = cur_node
+            output[checkpoint_name][cur_node_id] = cur_node
     f.close()
     return output
